% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_test_241007.R
\name{cat_test_241007}
\alias{cat_test_241007}
\title{Categorize trauma data}
\usage{
cat_test_241007(df, dx_pre, mech = TRUE, messages = FALSE, verbose = FALSE)
}
\arguments{
\item{df}{A dataframe in wide format containing ICD-10 diagnosis codes with a common column name prefix.
Diagnosis codes should be character strings and may have a decimal or not.}

\item{dx_pre}{Prefix for diagnosis code column names (example: dx1, dx2, etc.)}

\item{mech}{Should the program determine mechanism? Must be TRUE (default) or FALSE.
\itemize{
\item TRUE - Mechanism codes will be sought and processed by the program.
\item FALSE - This portion of the program will be skipped (to save time).
}}

\item{messages}{Should the program report completion of each step? Must be TRUE or FALSE (default).
\itemize{
\item TRUE - Messages will report completion of each step (may be helpful for large data sets).
\item FALSE - Messages will not be reported.
}}

\item{verbose}{Should other updates be printed to the console? TRUE or FALSE (default). This can be helpful for long running computations.}
}
\value{
A dataframe identical to the dataframe passed to the function with the following additional variables
         added:
         \itemize{
         \item sev_1-sev_n: AIS severity for diagnosis codes 1..n
         \item issbr_1-issbr_n: ISS body region for diagnosis codes 1..n
         \item mxaisbr1-mxaisbr6: maximum AIS severity for each of the 6 ISS body regions
         \item maxais: maximum AIS severity over all ISS body regions
         \item riss: computed injury severity score
         \item niss: new injury severity score
         \item ecode_1-ecode_4: first 4 mechanism codes found in each row of data
         \item mechmaj1-mechmaj4: CDC external cause of injury major mechanism for each E-Code captured
         \item intent1-intent4: intent for each E-Code captured
         \item lowmech: lowest CDC external cause of injury major mechanism for all E-Codes captured
         \item PmortTQIP: The TQIP model predicted probability of mortality
         \item PmortNIS: The NIS model predicted probability of mortality
         }
}
\description{
This function adds Abbreviated Injury Scores (AIS), Injury Severity Scores (ISS), and other descriptors of injury to a dataframe.
For each observation this function will
\enumerate{
   \item assign a severity (AIS) and ISS body region values to each valid ICD-10 injury diagnosis code,
   \item add variables for maximum severity of each body region,
   \item calculate ISS, "New ISS", maximum AIS, and regression-based mortality predictions,
   \item select the first 4 mechanism codes and categorize mechanism and intent following CDC guidelines
}
}
\details{
Data should be in wide format:
\tabular{rrrrr}{
ID    \tab \tab dx1      \tab \tab dx2     \tab\tab dx3    \tab\tab etc. \cr
31416 \tab\tab S32110A  \tab\tab S3251   \tab\tab NA     \tab\tab      \cr
31417 \tab\tab S72141A  \tab\tab T07XXXA \tab\tab D62    \tab\tab
}

Codes for AIS severity:
\itemize{
      \item 1 = Minor
      \item 2 = Moderate
      \item 3 = Serious
      \item 4 = Severe
      \item 5 = Critical
      \item 6 = Unsurvivable
      \item 9 = Unknown
}
}
\examples{
df_in <- read.table(header = TRUE, text = "
    ident   dx1       dx2       dx3
    31416   S32110A   S3251     NA
    31417   S72141A   T07XXXA   D62
")
df_out <- cat_trauma2(df_in, "dx", TRUE, FALSE)

}
